+++ packages/tailwindcss/src/index.test.ts
@@ -1292,6 +1292,215 @@ describe('Parsing themes values from CSS', () => {
       }"
     `)
   })
+
+  test('`default` theme values can be overridden by regular theme values`', async () => {
+    expect(
+      await compileCss(
+        css`
+          @theme {
+            --color-potato: #ac855b;
+          }
+          @theme default {
+            --color-potato: #efb46b;
+          }
+
+          @tailwind utilities;
+        `,
+        ['bg-potato'],
+      ),
+    ).toMatchInlineSnapshot(`
+      ":root {
+        --color-potato: #ac855b;
+      }
+
+      .bg-potato {
+        background-color: var(--color-potato, #ac855b);
+      }"
+    `)
+  })
+
+  test('`default` and `inline` can be used together', async () => {
+    expect(
+      await compileCss(
+        css`
+          @theme default inline {
+            --color-potato: #efb46b;
+          }
+
+          @tailwind utilities;
+        `,
+        ['bg-potato'],
+      ),
+    ).toMatchInlineSnapshot(`
+      ":root {
+        --color-potato: #efb46b;
+      }
+
+      .bg-potato {
+        background-color: #efb46b;
+      }"
+    `)
+  })
+
+  test('`default` and `reference` can be used together', async () => {
+    expect(
+      await compileCss(
+        css`
+          @theme default reference {
+            --color-potato: #efb46b;
+          }
+
+          @tailwind utilities;
+        `,
+        ['bg-potato'],
+      ),
+    ).toMatchInlineSnapshot(`
+      ".bg-potato {
+        background-color: var(--color-potato, #efb46b);
+      }"
+    `)
+  })
+
+  test('`default`, `inline`, and `reference` can be used together', async () => {
+    expect(
+      await compileCss(
+        css`
+          @theme default reference inline {
+            --color-potato: #efb46b;
+          }
+
+          @tailwind utilities;
+        `,
+        ['bg-potato'],
+      ),
+    ).toMatchInlineSnapshot(`
+      ".bg-potato {
+        background-color: #efb46b;
+      }"
+    `)
+  })
+
+  test('`default` can be used in `media(…)`', async () => {
+    expect(
+      await compileCss(
+        css`
+          @media theme() {
+            @theme {
+              --color-potato: #ac855b;
+            }
+          }
+          @media theme(default) {
+            @theme {
+              --color-potato: #efb46b;
+              --color-tomato: tomato;
+            }
+          }
+
+          @tailwind utilities;
+        `,
+        ['bg-potato', 'bg-tomato'],
+      ),
+    ).toMatchInlineSnapshot(`
+      ":root {
+        --color-potato: #ac855b;
+        --color-tomato: tomato;
+      }
+
+      .bg-potato {
+        background-color: var(--color-potato, #ac855b);
+      }
+
+      .bg-tomato {
+        background-color: var(--color-tomato, tomato);
+      }"
+    `)
+  })
+
+  test('`default` theme values can be overridden by plugin theme values', async () => {
+    let { build } = await compile(
+      css`
+        @theme default {
+          --color-red: red;
+        }
+        @theme {
+          --color-orange: orange;
+        }
+        @plugin "my-plugin";
+        @tailwind utilities;
+      `,
+      {
+        loadPlugin: async () => {
+          return plugin(({}) => {}, {
+            theme: {
+              extend: {
+                colors: {
+                  red: 'tomato',
+                  orange: '#f28500',
+                },
+              },
+            },
+          })
+        },
+      },
+    )
+
+    expect(optimizeCss(build(['text-red', 'text-orange'])).trim()).toMatchInlineSnapshot(`
+      ":root {
+        --color-orange: orange;
+      }
+
+      .text-orange {
+        color: var(--color-orange, orange);
+      }
+
+      .text-red {
+        color: tomato;
+      }"
+    `)
+  })
+
+  test('`default` theme values can be overridden by config theme values', async () => {
+    let { build } = await compile(
+      css`
+        @theme default {
+          --color-red: red;
+        }
+        @theme {
+          --color-orange: orange;
+        }
+        @config "./my-config.js";
+        @tailwind utilities;
+      `,
+      {
+        loadConfig: async () => {
+          return {
+            theme: {
+              extend: {
+                colors: {
+                  red: 'tomato',
+                  orange: '#f28500',
+                },
+              },
+            },
+          }
+        },
+      },
+    )
+
+    expect(optimizeCss(build(['text-red', 'text-orange'])).trim()).toMatchInlineSnapshot(`
+      ":root {
+        --color-orange: orange;
+      }
+
+      .text-orange {
+        color: var(--color-orange, orange);
+      }
+
+      .text-red {
+        color: tomato;
+      }"
+    `)
+  })
 })
 
 describe('plugins', () => {